New visuals

+++ vispy/scene/components/color.py
@@ -68,7 +68,7 @@ class VertexColorComponent(VisualComponent):
         super(VertexColorComponent, self).__init__()
         self._color = color
         self._vbo = None
-        
+
         # Create Varying to connect vertex / fragment shaders
         var = Varying('rgba', dtype='vec4')
         self._funcs['frag_color']['rgba'] = var
@@ -85,7 +85,7 @@ class VertexColorComponent(VisualComponent):
     @property
     def vbo(self):
         if self._vbo is None:
-            self._vbo = gloo.VertexBuffer(self._color)
+            self._vbo = gloo.VertexBuffer(self._color.astype(np.float32))
         return self._vbo
 
     def activate(self, program, mode):
